Demo/RL-DQN/train.py
import sys
import os
sys.path.append(os.path.dirname(os.path.abspath(__file__)))
sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
import torch
import yaml
from models.dqn_model import DQN, ReplayBuffer
from environments.cartpole_env import CartPoleEnvWrapper
from utils.logger import Logger

def train(config_path):
    # 加载配置
    with open(config_path) as f:
        config = yaml.safe_load(f)
    
    # 初始化环境和模型
    env = CartPoleEnvWrapper()
    policy_net = DQN(config['model']['input_size'], 
                    config['model']['hidden_size'],
                    config['model']['output_size'])
    target_net = DQN(config['model']['input_size'],
                    config['model']['hidden_size'],
                    config['model']['output_size'])
    target_net.load_state_dict(policy_net.state_dict())
    
    optimizer = torch.optim.Adam(policy_net.parameters())
    memory = ReplayBuffer(10000)
    
    # 训练循环
    for episode in range(config['training']['episodes']):
        state = env.reset()
        episode_reward = 0
        
        while True:
            # ε-贪婪策略选择动作
            if random.random() < epsilon:
                action = env.env.action_space.sample()
            else:
                with torch.no_grad():
                    action = policy_net(torch.FloatTensor(state)).argmax().item()
            
            # 环境交互
            next_state, reward, done, _ = env.step(action)
            memory.push(state, action, reward, next_state, done)
            
            # 网络更新
            if len(memory) > config['training']['batch_size']:
                transitions = memory.sample(config['training']['batch_size'])
                # ... 此处添加DQN训练逻辑 ...
            
            state = next_state
            episode_reward += reward
            
            if done:
                break

if __name__ == "__main__":
    import argparse
    parser = argparse.ArgumentParser()
    parser.add_argument("--config", type=str, required=True)
    args = parser.parse_args()
    train(args.config)