Demo/RL-DQN/requirements.txt
gym==0.26.2
torch==2.0.1
numpy>=1.26.0,<2.1.0  # 修正版本范围
pyyaml==6.0
tqdm==4.65.0
tensorflow==2.18.0    # 新增（如果需要）